requirements.txt
pyyaml>=6.0
anthropic>=0.50.0
pydantic>=2.0
pymysql>=1.1.0
tqdm>=4.60
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
